pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "A cli-utility to compare two files for differences"
authors = ["Arina Zubova <arina777pototskaya@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
flake8 = "^5.0.4"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"


[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
